remove old API folder and rename latest project.
