+++ BudgifyApi/BudgifyApi.sln
@@ -9,7 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgifyDal", "BudgifyDAL\Bu
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgifyModels", "BudgifyModels\BudgifyModels.csproj", "{F2A989D3-8218-4A0E-AE02-F1E16183C909}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgifyApi", "BudgifyApi3\BudgifyApi.csproj", "{D8F3B2D1-1CE1-42B8-A7D0-99C006DF4E79}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BudgifyApi", "BudgifyApi\BudgifyApi.csproj", "{E4651D02-D40E-4AF0-9914-66F9A162D2F7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +29,10 @@ Global
 		{F2A989D3-8218-4A0E-AE02-F1E16183C909}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F2A989D3-8218-4A0E-AE02-F1E16183C909}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F2A989D3-8218-4A0E-AE02-F1E16183C909}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D8F3B2D1-1CE1-42B8-A7D0-99C006DF4E79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D8F3B2D1-1CE1-42B8-A7D0-99C006DF4E79}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D8F3B2D1-1CE1-42B8-A7D0-99C006DF4E79}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D8F3B2D1-1CE1-42B8-A7D0-99C006DF4E79}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4651D02-D40E-4AF0-9914-66F9A162D2F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4651D02-D40E-4AF0-9914-66F9A162D2F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4651D02-D40E-4AF0-9914-66F9A162D2F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4651D02-D40E-4AF0-9914-66F9A162D2F7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
