+++ BudgifyApi/BudgifyApi/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "PostgreSQLConnection": "Server=144.22.175.160;Port=5432;Database=BudgifyDbTest;User Id=postgres;Password=budgifyPASS"
+    "Postgres": "Server=144.22.175.160;Port=5432;Database=BudgifyDbTest;User Id=postgres;Password=budgifyPASS"
   }
 }
