--- BudgifyApi/BudgifyApi3/Program.cs
@@ -1,60 +1,0 @@
-using BudgifyDal;
-using Microsoft.EntityFrameworkCore;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-var connectionString = builder.Configuration.GetConnectionString("Postgres");
-
-
-builder.Services.AddDbContext<AppDbContext>(options =>
-    options.UseNpgsql(connectionString));
-
-builder.Services.AddCors(options =>
-{
-    options.AddPolicy("NewPolicy", app =>
-    {
-        app.WithOrigins(
-            "http://localhost",
-            "http://localhost:4200",
-            "https://localhost:7230",
-            "http://localhost:90",
-            "http://localhost:3000",
-            "http://localhost:3000/login",
-            "http://localhost:3000/register")
-        .AllowAnyHeader()
-        .AllowAnyMethod()
-        .SetIsOriginAllowedToAllowWildcardSubdomains();
-
-
-    });
-});
-
-//builder.Services.AddDbContext<AppDbContext>(option => {
-//    option.UseNpgsql(builder.Configuration.GetConnectionString("Postgres"));
-//});
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseCors("NewPolicy");
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
