--- BudgifyApi/BudgifyApi3/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "Postgres": "Server=144.22.175.160;Port=5432;Database=BudgifyDbTest;User Id=postgres;Password=budgifyPASS"
-  }
-}
